requirements.txt
pandas==2.3.3
matplotlib==3.10.8
joblib==1.5.3
scikit-learn==1.8.0
sentence-transformers==5.3.0
python-dotenv==1.2.2
groq==1.1.1
fastapi==0.135.1
uvicorn==0.42.0
streamlit==1.55.0
requests==2.32.5